% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditr.R
\name{auditr}
\alias{auditr}
\title{R6 class: \code{auditr}}
\description{
\code{auditr} is an R6 class that provides methods for auditing two versions of stcopenmpp.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{args}}{A list of arguments passed into \code{auditr$new()}. See documentation for \code{auditr$new()} for details.}

\item{\code{code}}{A list of vectors representing code chunks. Any time an \code{auditr} method is called, the code chunk is saved and rendered in the audit report when \verb{$load_report()} is called.}

\item{\code{results}}{A list of four tibbles: \code{audit}, \code{output_tables}, \code{model_runs} and \code{performance}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-auditr-new}{\code{auditr$new()}}
\item \href{#method-auditr-load_releases}{\code{auditr$load_releases()}}
\item \href{#method-auditr-load_model}{\code{auditr$load_model()}}
\item \href{#method-auditr-build_models}{\code{auditr$build_models()}}
\item \href{#method-auditr-rebuild_openm}{\code{auditr$rebuild_openm()}}
\item \href{#method-auditr-load_output_tables}{\code{auditr$load_output_tables()}}
\item \href{#method-auditr-run_models}{\code{auditr$run_models()}}
\item \href{#method-auditr-load_report}{\code{auditr$load_report()}}
\item \href{#method-auditr-clone}{\code{auditr$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-auditr-new"></a>}}
\if{latex}{\out{\hypertarget{method-auditr-new}{}}}
\subsection{Method \code{new()}}{
This method creates an instance of an \code{auditr} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{auditr$new(
  releases,
  labels = paste0("Release ", 1:2),
  sln,
  mode = "release",
  bit = 64,
  git_url,
  git_user,
  git_token,
  git_commit,
  keep_vcxproj = TRUE,
  dir = getwd(),
  db,
  vs_cmd,
  platform_toolset,
  include,
  digits = 10
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{releases}}{Required: a length-two character vector representing the URLs or full paths to two releases of \code{stcopenmpp}.}

\item{\code{labels}}{Required: a length-two character vector representing the labels for the two releases (default: \code{paste0("Release ", 1:2)}).}

\item{\code{sln}}{Required: a length-one character vector representing the model solution file name.}

\item{\code{mode}}{Required: a length-one character vector representing the model mode to use during the model build (\code{release} or \code{debug}; default: \code{release}).}

\item{\code{bit}}{Required: a length-one numeric vector representing the the model architecture to use during the model build (\code{64} or \code{32}; default: \code{64}).}

\item{\code{git_url}}{Required: a length-one character vector representing the URL to the model repo.}

\item{\code{git_user}}{Optional: a length-one character vector representing the repo username.}

\item{\code{git_token}}{Optional: a length-one character vector representing the repo personal access token.}

\item{\code{git_commit}}{Optional: a length-one character vector representing commit or tag to checkout before building the model.}

\item{\code{keep_vcxproj}}{Optional: a length-one logical vector representing whether to keep the model's \code{Model.vcxproj} file when loaded or to discard and download a fresh template from the RiskPaths model (default: \code{TRUE}).}

\item{\code{dir}}{Optional: a length-one character vector representing the full path to where releases, models, model run meta and reports will be stored (default: \code{getwd()}).}

\item{\code{db}}{Optional: a length-one character vector representing the full path to where the model \code{.sqlite} file will be located for model runs.}

\item{\code{vs_cmd}}{Optional: a length-one character vector representing the full path to the Microsoft Visual Studio terminal. This is required if performing an audit in a Windows environment.}

\item{\code{platform_toolset}}{Optional: a length-one character vector representing the version of C++ build tools to use (e.g., \code{"v142"}, \code{"v143"}). This is only relevant for audits in Windows.}

\item{\code{include}}{Optional: a character vector representing the full path(s) to additional C++ libraries to include.}

\item{\code{digits}}{Optional: a length-one integer vector representing the number of digits to use when comparing output table values during audits.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns the \code{auditr} object invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-auditr-load_releases"></a>}}
\if{latex}{\out{\hypertarget{method-auditr-load_releases}{}}}
\subsection{Method \code{load_releases()}}{
This method downloads and extracts \code{stcopenmpp} archived releases (e.g., \code{.zip}, \code{.tar.gz}, \code{.tar.xz}).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{auditr$load_releases()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns the \code{auditr} object invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-auditr-load_model"></a>}}
\if{latex}{\out{\hypertarget{method-auditr-load_model}{}}}
\subsection{Method \code{load_model()}}{
This method downloads a model from a repo via \verb{git clone}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{auditr$load_model()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns the \code{auditr} object invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-auditr-build_models"></a>}}
\if{latex}{\out{\hypertarget{method-auditr-build_models}{}}}
\subsection{Method \code{build_models()}}{
This method compiles a model in both \code{stcopenmpp} releases asynchronously.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{auditr$build_models()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns the \code{auditr} object invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-auditr-rebuild_openm"></a>}}
\if{latex}{\out{\hypertarget{method-auditr-rebuild_openm}{}}}
\subsection{Method \code{rebuild_openm()}}{
This method re-compiles the OpenM++ library in both \code{stcopenmpp} releases asynchronously.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{auditr$rebuild_openm()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns the \code{auditr} object invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-auditr-load_output_tables"></a>}}
\if{latex}{\out{\hypertarget{method-auditr-load_output_tables}{}}}
\subsection{Method \code{load_output_tables()}}{
This method retrieves and stores a list of a model's non-hidden output tables via \code{dbcopy.exe}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{auditr$load_output_tables()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns the \code{auditr} object invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-auditr-run_models"></a>}}
\if{latex}{\out{\hypertarget{method-auditr-run_models}{}}}
\subsection{Method \code{run_models()}}{
This method initiates model runs in both \code{stcopenmpp} releases asynchronously.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{auditr$run_models(
  parameters = c(),
  sub_from = c(),
  cases = 1e+06,
  threads = 1,
  sub_samples = 1,
  param_dir = NULL,
  tables_per_run = 25
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parameters}}{Optional: a named vector (default: \code{c()}).}

\item{\code{sub_from}}{Optional: a named vector (default: \code{c()}).}

\item{\code{cases}}{Required: a length-one numeric vector representing the number of cases to run (default: \code{1e6}).}

\item{\code{threads}}{Required: a length-one numeric vector representing the number of threads to use during the model run (default: \code{1}).}

\item{\code{sub_samples}}{Required: a length-one numeric vector representing the number of sub-samples to use during the model run (default: \code{1}).}

\item{\code{param_dir}}{Optional: a length-one character vector representing the full path to the directory where the parameter files are located (default: \code{NULL}).}

\item{\code{tables_per_run}}{Optional: a length-one numeric vector representing the number of output tables per model run (default: \code{25}).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns the \code{auditr} object invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-auditr-load_report"></a>}}
\if{latex}{\out{\hypertarget{method-auditr-load_report}{}}}
\subsection{Method \code{load_report()}}{
This method renders an audit report in .html format.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{auditr$load_report(
  name = paste0("auditr-report-", self$args$model),
  dir = paste0(self$args$dir, "/auditr-reports")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Required: a length-one character vector representing the file name of the report (default: \code{paste0("auditr-report-", self$args$model)}).}

\item{\code{dir}}{Required: a length-one character vector representing the full path to the location where the report will be exported in \code{.html} format (default: \code{paste0(self$args$dir, "/auditr-reports")}).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns an .html-formatted report.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-auditr-clone"></a>}}
\if{latex}{\out{\hypertarget{method-auditr-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{auditr$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
