% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_plotter.R
\name{bubble_plotter}
\alias{bubble_plotter}
\title{Plot Plastic Pollution vs. GDP as a Bubble Chart}
\usage{
bubble_plotter(countries)
}
\arguments{
\item{countries}{A character vector of country names to include in the plot.}
}
\value{
A bubble plot showing total plastic pollution vs GDP where bubble
size is population.
}
\description{
Creates bubble charts showing the relationship between a country's
GDP and the amount of plastic pollution found per cleanup event, for the
years 2019 and 2020.
}
\examples{
bubble_plotter(c("United States of America", "Canada", "Nigeria"))
bubble_plotter("Taiwan")

}
