% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_probabilities.R
\name{names_probabilities}
\alias{names_probabilities}
\title{Classify surnames based on race and county demographics}
\usage{
names_probabilities(family, key)
}
\arguments{
\item{family}{A vector of family names (i.e. surnames or last names).}

\item{key}{A Census API Key. This is typically a long string of mixed-case letters and numbers.}
}
\value{
An object that classifies inputted surnames according to race and county demographics.
}
\description{
Returns an object that classifies inputted surnames according to race and county demographics.
}
\examples{
\dontrun{
family <- c("Chykina", "Crabtree", "Duck")
key <- "45b2kjsskd2335435kkmfdksmfkko"
names_probabilities(family, key)
}
}
\author{
Charles Crabtree \email{ccrabtr@umich.edu}
}
