% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_plotter.R
\name{density_plotter}
\alias{density_plotter}
\title{density_plotter}
\usage{
density_plotter(
  data,
  selected_year = 2019,
  cutoff = 3500,
  chosen_types = c("unknown", "hdpe", "ldpe", "other", "pet", "pp", "ps", "pvc")
)
}
\arguments{
\item{data}{a data frame with plastic audit numbers and gdp figures}

\item{selected_year}{optional, default 2019. 2019 or 2020 for our data}

\item{cutoff}{optional, default 3500. the cutoff for the density plot x axis. Can be useful to ignore large out liars or to zoom into the most common audit amounts}

\item{chosen_types}{optional, default includes: unknown, hdpe, ldpe, other, pet, pp, ps, pvc. what type of plastic is to be compared to eachother}
}
\value{
a plot
}
\description{
Plots a density ridge plot comparing the density of plastic type audit amounts to gdp based on user inputs
}
\examples{
density_plotter(load_data(),
                         selected_year = 2020,
                         cutoff = 2000,
                         chosen_types = c("hdpe", "pvc", "other"))
}
