% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plastic_proportions.R
\name{plastic_proportions}
\alias{plastic_proportions}
\title{Table of the proportions of each plastic type}
\usage{
plastic_proportions(countries = NULL, years = NULL)
}
\arguments{
\item{countries}{A character vector of one or more countries to be included in the table}

\item{years}{A numeric or character vector of one or two years to be included in the table}
}
\value{
A GT table with proportions.
}
\description{
This function creates a table displaying the proportion of each plastic type for a specified country and year.
}
