% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling.R
\name{rate_ify2}
\alias{rate_ify2}
\title{rate_ify2}
\usage{
rate_ify2(df, numerators, denominator)
}
\arguments{
\item{df}{a dataset with information to be rate-ified}

\item{numerators}{single column or vector of columns to be rate-ified. Tidy eval compatible}

\item{denominator}{single denominator column used to create the rate along with the numerators}
}
\value{
a data set with rate-ified columns
}
\description{
turns selected columns into rates based on a denominator column
}
\examples{
rate_ify2(df = load_data(), numerator = hdpe:pvc, denominator = total_waste)
}
