% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_analysis.R
\name{stat_analysis}
\alias{stat_analysis}
\title{stat_analysis}
\usage{
stat_analysis(
  fixed_vars = c("Population", "pop_density", "gdp", "year", "region")
)
}
\arguments{
\item{fixed_vars}{Character vector of fixed-effect predictor variables to
include in the mixed-effects model. Defaults to log population,
log population density, log GDP, year, and region.}
}
\value{
A formatted gt table containing the specified variables in the model
}
\description{
Fits a mixed-effects model analyzing log total plastic pollution based on
user-selected predictors with Country as a random effect and creates a
formatted Type III ANOVA table for the mixed-effects model
}
\examples{
stat_analysis(fixed_vars = c("Population", "year", "region"))
}
